refactor prismic client
